du thu

+++ src/pages/ProductManage.jsx
@@ -2,6 +2,7 @@
 import React, { Component } from "react";
 import ProductServices from "../services/ProductServices";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import axios from "axios";
 import {
   faPenToSquare,
   faEye,
@@ -15,8 +16,12 @@ class ProductManage extends Component {
     super(props);
     this.searchTimeout = null;
     this.state = {
+      products_search: [],
       products: [],
       search: "",
+      currentPage: 1,
+      recordPerPage: 10,
+      searchResult: "",
     };
     this.addProduct = this.addProduct.bind(this);
     this.editProduct = this.editProduct.bind(this);
@@ -25,14 +30,20 @@ class ProductManage extends Component {
   }
 
   deleteProduct(id) {
-    ProductServices.deleteProduct(id).then((res) => {
-      this.setState({
-        products: this.state.products.filter(
-          (product) => product.productId !== id
-        ),
+    const shouldDelete = window.confirm(
+      "Are you sure you want to delete this product?"
+    );
+
+    if (shouldDelete) {
+      ProductServices.deleteProduct(id).then((res) => {
+        this.setState({
+          products: this.state.products.filter(
+            (product) => product.productId !== id
+          ),
+        });
+        toast.success("Delete product successfully");
       });
-      toast.success("Delete product successfully");
-    });
+    }
   }
 
   editProduct(id) {
@@ -43,11 +54,68 @@ class ProductManage extends Component {
   }
 
   componentDidMount() {
-    ProductServices.getProducts().then((res) => {
-      this.setState({ products: res.data });
-      console.log(res.data);
+    ProductServices.getProducts_search().then((res) => {
+      this.setState({ products_search: res.data });
+      console.log("product search :", res.data);
     });
+    this.getBooksByPagination(this.state.currentPage);
   }
+
+  getBooksByPagination(currentPage) {
+    currentPage = currentPage - 1;
+    axios
+      .get(
+        "http://localhost:8080/pharmacy-online/admin/products/list-products?page=" +
+          currentPage +
+          "&size=" +
+          this.state.recordPerPage
+      )
+      .then((response) => response.data)
+      .then((data) => {
+        this.setState({
+          products: data.content,
+          totalPages: data.totalPages,
+          totalElements: data.totalElements,
+          currentPage: data.number + 1,
+        });
+      });
+  }
+
+  //Writing All the pagination functions
+  //Show Next page
+  showNextPage = () => {
+    if (
+      this.state.currentPage <
+      Math.ceil(this.state.totalElements / this.state.recordPerPage)
+    ) {
+      this.getBooksByPagination(this.state.currentPage + 1);
+    }
+  };
+  //Show Last Page
+  showLastPage = () => {
+    if (
+      this.state.currentPage <
+      Math.ceil(this.state.totalElements / this.state.recordPerPage)
+    ) {
+      this.getBooksByPagination(
+        Math.ceil(this.state.totalElements / this.state.recordPerPage)
+      );
+    }
+  };
+  //Show First page
+  showFirstPage = () => {
+    let firstPage = 1;
+    if (this.state.currentPage > firstPage) {
+      this.getBooksByPagination(firstPage);
+    }
+  };
+  //Show previous page
+  showPrevPage = () => {
+    let prevPage = 1;
+    if (this.state.currentPage > prevPage) {
+      this.getBooksByPagination(this.state.currentPage - prevPage);
+    }
+  };
 
   addProduct() {
     this.props.history.push("/add-product/_add");
@@ -60,12 +128,28 @@ class ProductManage extends Component {
     const searchTerm = event.target.value;
 
     this.searchTimeout = setTimeout(() => {
-      this.setState({ search: searchTerm.toLowerCase() });
-      console.log(searchTerm);
+      const searchValue = searchTerm.toLowerCase();
+      this.setState({
+        search: searchValue,
+        searchResult: searchValue
+          ? this.state.products_search.filter((product) =>
+              product.name.toLowerCase().includes(searchValue)
+            )
+          : [],
+      });
     }, 400);
   };
 
   render() {
+    const {
+      products,
+      currentPage,
+      totalPages,
+      recordPerPage,
+      searchResult,
+      search,
+    } = this.state;
+    const displayProducts = search ? searchResult : products;
     return (
       <div className="body-wrapper">
         <div class="container p-0">
@@ -100,41 +184,44 @@ class ProductManage extends Component {
                   </div>
                 </div>
               </div>
-              <table class="table table-hover">
-                <span></span>
+              <table className="table rounded table-hover shadow">
                 <thead>
                   <tr>
                     <th style={{ textAlign: "center" }}>Image</th>
                     <th>Name</th>
-                    <th>Brand</th>
                     <th>Price</th>
                     <th>Quantity</th>
-                    <th>Actions</th>
+                    <th>Brand</th>
+                    <th>Action</th>
                   </tr>
                 </thead>
                 <tbody>
-                  {this.state.products
-                    .filter((product) => {
-                      const searchTerm = this.state.search.toLowerCase();
-                      const productName = product.name.toLowerCase();
-                      return searchTerm === ""
-                        ? true
-                        : productName.includes(searchTerm);
-                    })
-                    .map((product) => (
-                      <tr key="{product.productId}">
+                  {displayProducts.length === 0 ? (
+                    <tr align="center">
+                      <td colSpan="5">No Record Found</td>
+                    </tr>
+                  ) : (
+                    displayProducts.map((product, index) => (
+                      <tr key={product.id}>
+                        {/* <td>{(recordPerPage * (currentPage - 1)) + index + 1}</td> */}
                         <td>
                           <img
                             style={{ width: "100%" }}
-                            src={product.imageUrls[0]}
+                            // src={`../assets/images/${product.imageUrls[0]}`}
+                            src={
+                              product.imageUrls[0]?.startsWith("https")
+                                ? product.imageUrls[0]
+                                : `assets/images/${product.imageUrls[0]}`
+                            }
                             alt=""
                             srcSet=""
                           />
                         </td>
                         <td>{product.name}</td>
-                        <td> {product.brand}</td>
                         <td>{product.price}</td>
                         <td>{product.quantity}</td>
+                        <td>{product.brand}</td>
+
                         <td>
                           <a
                             onClick={() => this.editProduct(product.productId)}
@@ -154,8 +241,6 @@ class ProductManage extends Component {
                           </a>
                           <a
                             type="button"
-                            data-toggle="modal"
-                            data-target={`#myModal${product.productId}`}
                             onClick={() =>
                               this.viewProductById(product.productId)
                             }
@@ -166,9 +251,70 @@ class ProductManage extends Component {
                           </a>
                         </td>
                       </tr>
-                    ))}
+                    ))
+                  )}
                 </tbody>
               </table>
+              {search === "" ? (
+                <div>
+                  <div
+                    style={{
+                      float: "left",
+                      fontFamily: "monospace",
+                      color: "#0275d8",
+                    }}
+                  >
+                    Page {currentPage} of {totalPages}
+                  </div>
+                  <div style={{ float: "right" }}>
+                    <div class="clearfix"></div>
+                    <nav aria-label="Page navigation example">
+                      <ul class="pagination">
+                        <li class="page-item">
+                          <a
+                            type="button"
+                            class="page-link"
+                            disabled={currentPage === 1 ? true : false}
+                            onClick={this.showPrevPage}
+                          >
+                            Previous
+                          </a>
+                        </li>
+                        <li class="page-item">
+                          <a
+                            type="button"
+                            class="page-link"
+                            disabled={currentPage === 1 ? true : false}
+                            onClick={this.showFirstPage}
+                          >
+                            First
+                          </a>
+                        </li>
+                        <li class="page-item">
+                          <a
+                            type="button"
+                            class="page-link"
+                            disabled={currentPage === totalPages ? true : false}
+                            onClick={this.showNextPage}
+                          >
+                            Next
+                          </a>
+                        </li>
+                        <li class="page-item">
+                          <a
+                            type="button"
+                            class="page-link"
+                            disabled={currentPage === totalPages ? true : false}
+                            onClick={this.showLastPage}
+                          >
+                            Last
+                          </a>
+                        </li>
+                      </ul>
+                    </nav>
+                  </div>
+                </div>
+              ) : null}
             </div>
           </div>
         </div>
